+++ libs/terraform/component/deployment-trigger/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/deployment-trigger-event.component';
+export * from './lib/deployment-trigger-role.component';
+export * from './lib/deployment-trigger.component';