+++ libs/terraform/group/pool/src/lib/configs/pool-group.config.ts
@@ -1,0 +1,7 @@
+import { PoolGroupDefaultConfig } from '../interfaces';
+
+export const POOL_GROUP_DEFAULT_CONFIG: PoolGroupDefaultConfig = {
+  subDomain: 'www',
+  reverseProxyConfig: {},
+  ownerPictureConfig: {},
+};