+++ libs/terraform/pool-group/src/lib/configs/pool-group.config.ts
@@ -1,9 +1,0 @@
-import { PoolGroupConfig } from '../interfaces';
-
-export const POOL_GROUP_DEFAULT_CONFIG: Pick<
-  PoolGroupConfig,
-  'subDomain' | 'reverseProxyDesiredCount'
-> = {
-  subDomain: 'www',
-  reverseProxyDesiredCount: 1,
-};