+++ libs/terraform/postgres-service/src/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/configs';
-export * from './lib/postgres.component';
-export * from './lib/postgres.container';