+++ libs/terraform/service/reverse-proxy/src/index.ts
@@ -1,0 +1,4 @@
+export * from './lib/interfaces';
+export * from './lib/reverse-proxy-task-role.component';
+export * from './lib/reverse-proxy.component';
+export * from './lib/reverse-proxy.container';