+++ libs/terraform/aws-stack/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.1](https://github.com/tractr/stack/compare/terraform-aws-stack-0.4.0...terraform-aws-stack-0.4.1) (2021-10-12)
+
+
+### Bug Fixes
+
+* **terraform:** force update ([#285](https://github.com/tractr/stack/issues/285)) ([78f1e4e](https://github.com/tractr/stack/commit/78f1e4e16119d6ca52dc6c6842775775391d45bb))
+
+
+
+# [0.4.0](https://github.com/tractr/stack/compare/terraform-aws-stack-0.3.0...terraform-aws-stack-0.4.0) (2021-10-12)
+
+
+### Features
+
+* terraform handles multiple environement ([#283](https://github.com/tractr/stack/issues/283)) ([43d6e65](https://github.com/tractr/stack/commit/43d6e65dc348651db09a278fd000aabdedcbebec))
+
+
+
 # [0.3.0](https://github.com/tractr/stack/compare/terraform-aws-stack-0.2.1...terraform-aws-stack-0.3.0) (2021-09-01)